// src/generated/models/NewRegistrodevendaModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export interface NewRegistrodevendaBase {
  new_registrodevendaid: string;
  new_categoriadoproduto?: string;
  new_custodecompratotal?: number;
  new_datadavenda?: string;
  new_descricaodoproduto?: string;
  new_id?: string;
  new_lucrobrutototal?: number;
  new_nomedoarquiteto?: string;
  new_nomedocliente?: string;
  new_nomedofabricante?: string;
  new_nomedovendedor?: string;
  new_quantidade?: number;
  new_referenciadoprodutomodelo?: string;
  new_valordeprodutototal?: number;
  new_valordeservicototal?: number;
  new_valortotal?: number;
  "new_ItemdeOrcamento@odata.bind"?: string;
  "new_Orcamento@odata.bind"?: string;
  "new_PrecodeProduto@odata.bind"?: string;
  "new_Produto@odata.bind"?: string;
  "new_Projeto@odata.bind"?: string;
  statecode?: number;
  statuscode?: number;
}

export interface NewRegistrodevenda extends NewRegistrodevendaBase {
  _new_itemdeorcamento_value?: string;
  _new_orcamento_value?: string;
  _new_precodeproduto_value?: string;
  _new_produto_value?: string;
  _new_projeto_value?: string;
}